Chore: move editor text elements into separate files when possible
