+++ lib/src/editor/raw_editor/raw_editor_state.dart
@@ -32,9 +32,10 @@ import '../widgets/default_styles.dart';
 import '../widgets/keyboard_service_widget.dart';
 import '../widgets/link.dart';
 import '../widgets/proxy.dart';
-import '../widgets/text/text_block.dart';
-import '../widgets/text/text_line.dart';
-import '../widgets/text/text_selection.dart';
+import '../widgets/text/block_line/text_block.dart';
+import '../widgets/text/line/editable_text_line.dart';
+import '../widgets/text/line/text_line.dart';
+import '../widgets/text/selection/text_selection.dart';
 import 'raw_editor.dart';
 import 'raw_editor_actions.dart';
 import 'raw_editor_render_object.dart';
