+++ lib/src/editor/widgets/text/line/render_text_line_element.dart
@@ -1,0 +1,114 @@
+import 'package:flutter/material.dart'
+    show
+        Element,
+        ElementVisitor,
+        RenderBox,
+        RenderObject,
+        RenderObjectElement,
+        Widget;
+import 'package:meta/meta.dart' show internal;
+import '../../box.dart' show RenderContentProxyBox;
+import 'editable_text_line.dart' show EditableTextLine;
+import 'render_editable_text_line.dart' show RenderEditableTextLine;
+import 'text_line.dart' show TextLineSlot;
+
+//TODO: we need to document what does this element
+// and why we use it
+@internal
+class TextLineElement extends RenderObjectElement {
+  TextLineElement(EditableTextLine super.line);
+
+  final Map<TextLineSlot, Element> _slotToChildren = <TextLineSlot, Element>{};
+
+  @override
+  EditableTextLine get widget => super.widget as EditableTextLine;
+
+  @override
+  RenderEditableTextLine get renderObject =>
+      super.renderObject as RenderEditableTextLine;
+
+  @override
+  void visitChildren(ElementVisitor visitor) {
+    _slotToChildren.values.forEach(visitor);
+  }
+
+  @override
+  void forgetChild(Element child) {
+    assert(_slotToChildren.containsValue(child));
+    assert(child.slot is TextLineSlot);
+    assert(_slotToChildren.containsKey(child.slot));
+    _slotToChildren.remove(child.slot);
+    super.forgetChild(child);
+  }
+
+  @override
+  void mount(Element? parent, dynamic newSlot) {
+    super.mount(parent, newSlot);
+    _mountChild(widget.leading, TextLineSlot.leading);
+    _mountChild(widget.body, TextLineSlot.body);
+  }
+
+  @override
+  void update(EditableTextLine newWidget) {
+    super.update(newWidget);
+    assert(widget == newWidget);
+    _updateChild(widget.leading, TextLineSlot.leading);
+    _updateChild(widget.body, TextLineSlot.body);
+  }
+
+  @override
+  void insertRenderObjectChild(RenderBox child, TextLineSlot? slot) {
+    // assert(child is RenderBox);
+    _updateRenderObject(child, slot);
+    assert(renderObject.children.keys.contains(slot));
+  }
+
+  @override
+  void removeRenderObjectChild(RenderObject child, TextLineSlot? slot) {
+    assert(child is RenderBox);
+    assert(renderObject.children[slot!] == child);
+    _updateRenderObject(null, slot);
+    assert(!renderObject.children.keys.contains(slot));
+  }
+
+  @override
+  void moveRenderObjectChild(
+      RenderObject child, dynamic oldSlot, dynamic newSlot) {
+    throw UnimplementedError();
+  }
+
+  void _mountChild(Widget? widget, TextLineSlot slot) {
+    final oldChild = _slotToChildren[slot];
+    final newChild = updateChild(oldChild, widget, slot);
+    if (oldChild != null) {
+      _slotToChildren.remove(slot);
+    }
+    if (newChild != null) {
+      _slotToChildren[slot] = newChild;
+    }
+  }
+
+  void _updateRenderObject(RenderBox? child, TextLineSlot? slot) {
+    switch (slot) {
+      case TextLineSlot.leading:
+        renderObject.setLeading(child);
+        break;
+      case TextLineSlot.body:
+        renderObject.setBody(child as RenderContentProxyBox?);
+        break;
+      default:
+        throw UnimplementedError();
+    }
+  }
+
+  void _updateChild(Widget? widget, TextLineSlot slot) {
+    final oldChild = _slotToChildren[slot];
+    final newChild = updateChild(oldChild, widget, slot);
+    if (oldChild != null) {
+      _slotToChildren.remove(slot);
+    }
+    if (newChild != null) {
+      _slotToChildren[slot] = newChild;
+    }
+  }
+}
